Remove old usages implementation
