+++ plugins/plugin-java/che-plugin-java-ext-lang-client/src/main/java/org/eclipse/che/ide/ext/java/client/inject/JavaGinModule.java
@@ -51,9 +51,6 @@ import org.eclipse.che.ide.ext.java.client.reference.JavaFqnProvider;
 import org.eclipse.che.ide.ext.java.client.resource.ClassInterceptor;
 import org.eclipse.che.ide.ext.java.client.resource.JavaSourceRenameValidator;
 import org.eclipse.che.ide.ext.java.client.resource.SourceFolderInterceptor;
-import org.eclipse.che.ide.ext.java.client.search.JavaSearchJsonRpcClient;
-import org.eclipse.che.ide.ext.java.client.search.JavaSearchService;
-import org.eclipse.che.ide.ext.java.client.search.node.NodeFactory;
 import org.eclipse.che.ide.ext.java.client.settings.compiler.ErrorsWarningsPreferenceManager;
 import org.eclipse.che.ide.ext.java.client.settings.compiler.JavaCompilerPreferenceManager;
 import org.eclipse.che.ide.ext.java.client.settings.compiler.JavaCompilerPreferencePresenter;
@@ -86,7 +83,6 @@ public class JavaGinModule extends AbstractGinModule {
     bind(NewJavaSourceFileView.class).to(NewJavaSourceFileViewImpl.class).in(Singleton.class);
     bind(QuickDocumentation.class).to(QuickDocPresenter.class).in(Singleton.class);
     bind(JavaNavigationService.class).to(JavaNavigationServiceImpl.class);
-    bind(JavaSearchService.class).to(JavaSearchJsonRpcClient.class);
 
     GinMultibinder.newSetBinder(binder(), NodeInterceptor.class)
         .addBinding()
@@ -121,7 +117,6 @@ public class JavaGinModule extends AbstractGinModule {
             .implement(PropertyWidget.class, PropertyWidgetImpl.class)
             .build(PropertyWidgetFactory.class));
 
-    install(new GinFactoryModuleBuilder().build(NodeFactory.class));
     install(
         new GinFactoryModuleBuilder()
             .build(org.eclipse.che.ide.ext.java.client.navigation.factory.NodeFactory.class));
