+++ plugins/plugin-java/che-plugin-java-ext-lang-client/src/main/java/org/eclipse/che/ide/ext/java/client/search/FindUsagesPresenter.java
@@ -5,44 +5,47 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
+ * Contributors: Red Hat, Inc. - initial API and implementation
  */
 package org.eclipse.che.ide.ext.java.client.search;
 
 import static org.eclipse.che.ide.api.notification.StatusNotification.DisplayMode.FLOAT_MODE;
 import static org.eclipse.che.ide.api.notification.StatusNotification.Status.FAIL;
 
-import com.google.common.base.Optional;
-import com.google.gwt.json.client.JSONParser;
 import com.google.gwt.user.client.ui.AcceptsOneWidget;
 import com.google.gwt.user.client.ui.IsWidget;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import org.eclipse.che.api.promises.client.Operation;
-import org.eclipse.che.api.promises.client.OperationException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import org.eclipse.che.api.languageserver.shared.model.SnippetParameters;
+import org.eclipse.che.api.languageserver.shared.model.SnippetResult;
+import org.eclipse.che.api.promises.client.Function;
 import org.eclipse.che.api.promises.client.Promise;
-import org.eclipse.che.api.promises.client.PromiseError;
+import org.eclipse.che.api.promises.client.PromiseProvider;
 import org.eclipse.che.ide.Resources;
+import org.eclipse.che.ide.api.editor.document.Document;
 import org.eclipse.che.ide.api.editor.texteditor.TextEditor;
 import org.eclipse.che.ide.api.notification.NotificationManager;
 import org.eclipse.che.ide.api.parts.PartStackType;
 import org.eclipse.che.ide.api.parts.WorkspaceAgent;
+import org.eclipse.che.ide.api.parts.base.BaseActionDelegate;
 import org.eclipse.che.ide.api.parts.base.BasePresenter;
-import org.eclipse.che.ide.api.resources.Container;
-import org.eclipse.che.ide.api.resources.Project;
-import org.eclipse.che.ide.api.resources.Resource;
-import org.eclipse.che.ide.api.resources.VirtualFile;
-import org.eclipse.che.ide.commons.exception.ServerException;
-import org.eclipse.che.ide.dto.DtoFactory;
 import org.eclipse.che.ide.ext.java.client.JavaLocalizationConstant;
-import org.eclipse.che.ide.ext.java.client.resource.SourceFolderMarker;
-import org.eclipse.che.ide.ext.java.client.util.JavaUtil;
-import org.eclipse.che.ide.ext.java.shared.dto.search.FindUsagesRequest;
-import org.eclipse.che.ide.ext.java.shared.dto.search.FindUsagesResponse;
-import org.eclipse.che.ide.rest.HTTPStatus;
+import org.eclipse.che.ide.ext.java.client.service.JavaLanguageExtensionServiceClient;
+import org.eclipse.che.ide.ui.smartTree.data.Node;
 import org.eclipse.che.ide.util.loging.Log;
+import org.eclipse.che.jdt.ls.extension.api.dto.LinearRange;
+import org.eclipse.che.jdt.ls.extension.api.dto.SearchResult;
+import org.eclipse.che.jdt.ls.extension.api.dto.UsagesResponse;
+import org.eclipse.che.plugin.languageserver.ide.editor.LanguageServerEditorConfiguration;
+import org.eclipse.che.plugin.languageserver.ide.service.TextDocumentServiceClient;
+import org.eclipse.che.plugin.languageserver.ide.util.DtoBuildHelper;
 import org.eclipse.che.providers.DynaObject;
+import org.eclipse.lsp4j.TextDocumentPositionParams;
 import org.vectomatic.dom.svg.ui.SVGResource;
 
 /**
@@ -52,32 +55,42 @@ import org.vectomatic.dom.svg.ui.SVGResource;
  */
 @Singleton
 @DynaObject
-public class FindUsagesPresenter extends BasePresenter implements FindUsagesView.ActionDelegate {
+public class FindUsagesPresenter extends BasePresenter implements BaseActionDelegate {
 
   private WorkspaceAgent workspaceAgent;
   private JavaLocalizationConstant localizationConstant;
   private FindUsagesView view;
-  private JavaSearchService searchService;
-  private DtoFactory dtoFactory;
+  private JavaLanguageExtensionServiceClient searchService;
+  private TextDocumentServiceClient textDocumentService;
   private NotificationManager manager;
   private final Resources resources;
+  private DtoBuildHelper dtoBuildHelper;
+  private Map<String, Map<LinearRange, SnippetResult>> cache = new HashMap<>();
+  private PromiseProvider promiseProvider;
+  private NodeFactory nodeFactory;
 
   @Inject
   public FindUsagesPresenter(
       WorkspaceAgent workspaceAgent,
       JavaLocalizationConstant localizationConstant,
       FindUsagesView view,
-      JavaSearchService searchService,
-      DtoFactory dtoFactory,
+      JavaLanguageExtensionServiceClient searchService,
+      TextDocumentServiceClient textDocumentService,
+      DtoBuildHelper dtoBuildHelper,
       NotificationManager manager,
-      Resources resources) {
+      Resources resources,
+      PromiseProvider promiseProvider,
+      NodeFactory nodeFactory) {
     this.workspaceAgent = workspaceAgent;
     this.localizationConstant = localizationConstant;
     this.view = view;
     this.searchService = searchService;
-    this.dtoFactory = dtoFactory;
+    this.textDocumentService = textDocumentService;
+    this.dtoBuildHelper = dtoBuildHelper;
     this.manager = manager;
     this.resources = resources;
+    this.promiseProvider = promiseProvider;
+    this.nodeFactory = nodeFactory;
     view.setDelegate(this);
   }
 
@@ -106,85 +119,97 @@ public class FindUsagesPresenter extends BasePresenter implements FindUsagesView
     container.setWidget(view);
   }
 
-  public void findUsages(TextEditor activeEditor) {
-
-    final VirtualFile virtualFile = activeEditor.getEditorInput().getFile();
-
-    if (virtualFile instanceof Resource) {
-      final Project project = ((Resource) virtualFile).getRelatedProject().get();
-
-      if (project == null) {
-        return;
-      }
-
-      final Optional<Resource> srcFolder =
-          ((Resource) virtualFile).getParentWithMarker(SourceFolderMarker.ID);
-
-      if (!srcFolder.isPresent()) {
-        return;
-      }
-
-      final String fqn = JavaUtil.resolveFQN((Container) srcFolder.get(), (Resource) virtualFile);
-
-      String projectPath = project.getLocation().toString();
-      FindUsagesRequest request = dtoFactory.createDto(FindUsagesRequest.class);
-      request.setFQN(fqn);
-      request.setProjectPath(projectPath);
-      request.setOffset(activeEditor.getCursorOffset());
-
-      Promise<FindUsagesResponse> promise = searchService.findUsages(request);
-      promise
-          .then(
-              new Operation<FindUsagesResponse>() {
-                @Override
-                public void apply(FindUsagesResponse arg) throws OperationException {
-                  handleResponse(arg);
-                }
-              })
-          .catchError(
-              new Operation<PromiseError>() {
-                @Override
-                public void apply(PromiseError arg) throws OperationException {
-                  Throwable cause = arg.getCause();
-                  if (cause instanceof ServerException) {
-                    handleError(((ServerException) cause).getHTTPStatus(), cause.getMessage());
-                    return;
-                  }
-                  // in case websocket request
-                  if (cause
-                      instanceof org.eclipse.che.ide.websocket.rest.exceptions.ServerException) {
-                    handleError(
-                        ((org.eclipse.che.ide.websocket.rest.exceptions.ServerException) cause)
-                            .getHTTPStatus(),
-                        cause.getMessage());
-                    return;
-                  }
-                  Log.error(getClass(), arg);
-                  manager.notify(
-                      localizationConstant.failedToProcessFindUsage(),
-                      arg.getMessage(),
-                      FAIL,
-                      FLOAT_MODE);
-                }
-              });
+  public void findUsages(TextEditor editor) {
+    if (!(editor.getConfiguration() instanceof LanguageServerEditorConfiguration)) {
+      return;
     }
+
+    Document document = editor.getDocument();
+    TextDocumentPositionParams paramsDTO =
+        dtoBuildHelper.createTDPP(document, editor.getCursorOffset());
+
+    Promise<UsagesResponse> promise = searchService.usages(paramsDTO);
+    promise
+        .then(this::handleResponse)
+        .catchError(
+            arg -> {
+              Throwable cause = arg.getCause();
+              Log.error(getClass(), cause);
+              manager.notify(
+                  localizationConstant.failedToProcessFindUsage(),
+                  arg.getMessage(),
+                  FAIL,
+                  FLOAT_MODE);
+            });
   }
 
-  private void handleError(int statusCode, String message) {
-    if (statusCode == HTTPStatus.BAD_REQUEST) {
-      manager.notify(
-          localizationConstant.failedToProcessFindUsage(),
-          JSONParser.parseLenient(message).isObject().get("message").isString().stringValue(),
-          FAIL,
-          FLOAT_MODE);
-    } else {
-      manager.notify(localizationConstant.failedToProcessFindUsage(), message, FAIL, FLOAT_MODE);
-    }
-  }
-
-  private void handleResponse(FindUsagesResponse response) {
+  private void handleResponse(UsagesResponse response) {
     workspaceAgent.openPart(this, PartStackType.INFORMATION);
     workspaceAgent.setActivePart(this);
+    cache.clear();
     view.showUsages(response);
+  }
+
+  public Promise<List<Node>> computeMatches(ElementNode elementNode) {
+    Node ancestor = elementNode.getParent();
+    SearchResult rootElement = elementNode.getElement();
+
+    while ((ancestor instanceof ElementNode)
+        && ((ElementNode) ancestor)
+            .getElement()
+            .getUri()
+            .equals(elementNode.getElement().getUri())) {
+      rootElement = ((ElementNode) ancestor).getElement();
+      ancestor = ancestor.getParent();
+    }
+
+    return loadMatches(rootElement)
+        .then(
+            (Function<Map<LinearRange, SnippetResult>, List<Node>>)
+                matches -> {
+                  return filterMatches(elementNode.getElement(), matches);
+                });
+  }
+
+  private List<Node> filterMatches(SearchResult sr, Map<LinearRange, SnippetResult> matches) {
+    return sr.getMatches()
+        .stream()
+        .filter(
+            range -> {
+              return matches.containsKey(range);
+            })
+        .map(range -> nodeFactory.createMatch(sr.getUri(), matches.get(range)))
+        .collect(Collectors.toList());
+  }
+
+  private Promise<Map<LinearRange, SnippetResult>> loadMatches(SearchResult element) {
+    Map<LinearRange, SnippetResult> snippets = cache.get(element.getUri());
+    if (snippets != null) {
+      return promiseProvider.resolve(snippets);
+    }
+
+    List<LinearRange> ranges = new ArrayList<>();
+    collectMatchRanges(element, ranges);
+    SnippetParameters params =
+        new org.eclipse.che.api.languageserver.shared.dto.DtoClientImpls.SnippetParametersDto(
+            new SnippetParameters(element.getUri(), ranges));
+
+    return textDocumentService
+        .getSnippets(params)
+        .then(
+            (List<SnippetResult> results) -> {
+              Map<LinearRange, SnippetResult> snippets2 = new HashMap<>();
+              cache.put(element.getUri(), snippets2);
+              results.forEach(
+                  result -> {
+                    snippets2.put(result.getLinearRange(), result);
+                  });
+              return snippets2;
+            });
+  }
+
+  private void collectMatchRanges(SearchResult element, List<LinearRange> ranges) {
+    ranges.addAll(element.getMatches());
+    element.getChildren().forEach(child -> collectMatchRanges(child, ranges));
   }
 }
