+++ plugins/plugin-java/che-plugin-java-ext-lang-server/src/main/java/org/eclipse/che/plugin/java/server/rest/JdtExceptionMapper.java
@@ -5,8 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
+ * Contributors: Red Hat, Inc. - initial API and implementation
  */
 package org.eclipse.che.plugin.java.server.rest;
 
@@ -17,7 +16,6 @@ import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
 import org.eclipse.che.dto.server.DtoFactory;
 import org.eclipse.che.plugin.java.server.JdtException;
-import org.eclipse.che.plugin.java.server.search.SearchException;
 
 /**
  * Exception mapper for all JDT exceptions
@@ -29,15 +27,9 @@ import org.eclipse.che.plugin.java.server.search.SearchException;
 public class JdtExceptionMapper implements ExceptionMapper<JdtException> {
   @Override
   public Response toResponse(JdtException exception) {
-    if (exception instanceof SearchException) {
-      return Response.status(Response.Status.BAD_REQUEST)
-          .entity(DtoFactory.getInstance().toJson(exception.getServiceError()))
-          .type(MediaType.APPLICATION_JSON)
-          .build();
-    } else
-      return Response.serverError()
-          .entity(DtoFactory.getInstance().toJson(exception.getServiceError()))
-          .type(MediaType.APPLICATION_JSON)
-          .build();
+    return Response.serverError()
+        .entity(DtoFactory.getInstance().toJson(exception.getServiceError()))
+        .type(MediaType.APPLICATION_JSON)
+        .build();
   }
 }
